pyproject.toml
[tool.poetry]
name = "semantic-search-demo"
version = "0.1.0"
description = ""
authors = ["eolseng <e.olseng@gmail.com>"]
readme = "README.md"
packages = [{include = "semantic_search_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
chromadb = "^0.4.24"
openai = "^1.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
